requirements.txt
Jinja2
jingo
pygments
zope.structuredtext
redis
hiredis
Celery
django-celery
akismet
django-celery-email
isodate
django-redis-cache
django_compressor
django-bootstrap-form
markdown
